requirements.txt
requests >= 2.26.0
web3 >= 5.24.0
prometheus-client >= 0.12.0
pyyaml >= 6.0
pymongo >= 3.12.1
discord_webhook >= 0.14.0
